Create table_model
This module contains classes for creating and interacting with a model-view table system, including an abstract table model, sort proxy filter model, and one or more table views. It allows table data, objects, and operations to be housed within a single class. Data currently need to be in the form of a list of column names, a list of lists containing row data, and a list of dictionaries containing details about the table columns (e.g. specified column width).